Update README.md
This is a small tool which was made with the help of python language with the help of nlp which is known as natural language process. The tool is used to find the rating or public response of any particular product or celebrity. The result can be either positive, negative or neutral.